Вынесла логику в отдельный файл
