+++ js/test/timer.test.js
